STY: move project_hwm to pyglow methods
