+++ pysatMissionPlanning/instruments/pysat_sgp4.py
@@ -63,7 +63,7 @@ def init(self):
                     'B_sc_z', meta=[in_meta.copy(), in_meta.copy(),
                                     in_meta.copy()])
     # project total wind vector
-    self.custom.add(project_hwm_onto_sc, 'modify')
+    self.custom.add(methglow.project_hwm_onto_sc, 'modify')
 
 
 def load(fnames, tag=None, sat_id=None, obs_long=0., obs_lat=0., obs_alt=0.,
@@ -283,41 +283,3 @@ def download(date_array, tag, sat_id, data_path=None, user=None,
     """ Data is simulated so no download routine is possible. Simple pass
     function"""
     pass
-
-
-def project_hwm_onto_sc(inst):
-
-    import pysatMissionPlanning.methods.attitude as methatt
-
-    def get_wind_comp(inst, direction='x'):
-        unit_zon = 'unit_zonal_wind_ecef_' + direction
-        unit_mer = 'unit_mer_wind_ecef_' + direction
-
-        return (inst['zonal_wind']*inst[unit_zon] +
-                inst['meridional_wind']*inst[unit_mer])
-
-    inst['total_wind_x'] = get_wind_comp(inst, direction='x')
-    inst['total_wind_y'] = get_wind_comp(inst, direction='y')
-    inst['total_wind_z'] = get_wind_comp(inst, direction='z')
-
-    methatt.project_ecef_vector_onto_sc(inst, 'total_wind_x', 'total_wind_y',
-                                        'total_wind_z', 'sim_wind_sc_x',
-                                        'sim_wind_sc_y', 'sim_wind_sc_z')
-
-    inst.meta['sim_wind_sc_x'] = {'units': 'm/s',
-                                  'long_name': 'Simulated x-vector ' +
-                                  'instrument wind',
-                                  'desc': 'Wind from model as measured ' +
-                                  'by instrument in its x-direction'}
-    inst.meta['sim_wind_sc_y'] = {'units': 'm/s',
-                                  'long_name': 'Simulated y-vector ' +
-                                  'instrument wind',
-                                  'desc': 'Wind from model as measured ' +
-                                  'by instrument in its y-direction'}
-    inst.meta['sim_wind_sc_z'] = {'units': 'm/s',
-                                  'long_name': 'Simulated z-vector ' +
-                                  'instrument wind',
-                                  'desc': 'Wind from model as measured ' +
-                                  'by instrument in its z-direction'}
-
-    return
